Fix socket file structure
